+++ src/main/kotlin/kr/hhplus/be/server/point/domain/UserPointHistoryRepository.kt
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.point.domain
+
+interface UserPointHistoryRepository {
+    fun findAllByUserId(userId: Long): List<UserPointHistory>
+    fun save(pointHistory: UserPointHistory): UserPointHistory
+}
