+++ src/main/kotlin/kr/hhplus/be/server/point/domain/UserPointRepository.kt
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.point.domain
+
+interface UserPointRepository {
+    fun getByUserId(userId: Long): UserPoint
+    fun save(userPoint: UserPoint): UserPoint
+}
