--- src/main/kotlin/kr/hhplus/be/server/user/domain/UserPointHistoryRepository.kt
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.user.domain
-
-interface UserPointHistoryRepository {
-    fun findAllByUserId(userId: Long): List<UserPointHistory>
-    fun save(pointHistory: UserPointHistory): UserPointHistory
-}
