--- src/main/kotlin/kr/hhplus/be/server/user/domain/UserPointRepository.kt
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.user.domain
-
-interface UserPointRepository {
-    fun getByUserId(userId: Long): UserPoint
-    fun save(userPoint: UserPoint): UserPoint
-}
