--- src/main/kotlin/kr/hhplus/be/server/user/infrastructure/DefaultUserPointRepository.kt
@@ -1,17 +1,0 @@
-package kr.hhplus.be.server.user.infrastructure
-
-import kr.hhplus.be.server.shared.exception.ResourceNotFoundException
-import kr.hhplus.be.server.user.domain.UserPoint
-import kr.hhplus.be.server.user.domain.UserPointRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-class DefaultUserPointRepository(private val jpaRepository: JpaUserPointRepository): UserPointRepository {
-    override fun getByUserId(userId: Long): UserPoint {
-        return jpaRepository.findByUserId(userId) ?: throw ResourceNotFoundException()
-    }
-
-    override fun save(userPoint: UserPoint): UserPoint {
-        return jpaRepository.save(userPoint)
-    }
-}
