--- src/main/kotlin/kr/hhplus/be/server/user/infrastructure/JpaUserPointRepository.kt
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.user.infrastructure
-
-import kr.hhplus.be.server.user.domain.UserPoint
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface JpaUserPointRepository: JpaRepository<UserPoint, Long> {
-    fun findByUserId(userId: Long): UserPoint?
-}
